move files related to dns encoding to the dns directory
