+++ packages/lib/src/offchainResolver/encoding/decodeCalldata.test.ts
@@ -1,6 +1,8 @@
 import { ethers } from 'ethers';
 import { decodeCalldata } from './decodeCalldata';
 import { getResolverInterface } from './getResolverInterface';
+import { encodeEnsName } from '../dns/encodeEnsName';
+
 describe('decodeCalldata', () => {
     it('decodes valid calldata', () => {
         const calldata =
@@ -21,7 +23,7 @@ describe('decodeCalldata', () => {
         ]);
 
         const calldata = getResolverInterface().encodeFunctionData('resolve', [
-            dnsName('foo.dm3.eth'),
+            encodeEnsName('foo.dm3.eth'),
             textData,
         ]);
 
@@ -30,28 +32,3 @@ describe('decodeCalldata', () => {
         );
     });
 });
-function dnsName(name: string) {
-    // strip leading and trailing .
-    const n = name.replace(/^\.|\.$/gm, '');
-
-    var bufLen = n === '' ? 1 : n.length + 2;
-    var buf = Buffer.allocUnsafe(bufLen);
-
-    let offset = 0;
-    if (n.length) {
-        const list = n.split('.');
-        for (let i = 0; i < list.length; i++) {
-            const len = buf.write(list[i], offset + 1);
-            buf[offset] = len;
-            offset += len + 1;
-        }
-    }
-    buf[offset++] = 0;
-    return (
-        '0x' +
-        buf.reduce(
-            (output, elem) => output + ('0' + elem.toString(16)).slice(-2),
-            '',
-        )
-    );
-}
