+++ packages/lib/src/offchainResolver/index.ts
@@ -1,4 +1,6 @@
 export { decodeCalldata } from './encoding/decodeCalldata';
 export { encodeUserProfile } from './encoding/encodeUserProfile';
+export { decodeDnsName } from './dns/decodeDnsName';
+export { encodeEnsName } from './dns/encodeEnsName';
 
 export type { DecodedCcipRequest } from './types';
